+++ test/e2e/managedclustersetbinding_test.go
@@ -44,6 +44,11 @@ var _ = ginkgo.Describe("ManagedClusterSetBinding", func() {
 		})
 
 		ginkgo.AfterEach(func() {
+			if ginkgo.CurrentSpecReport().Failed() {
+				ginkgo.By(fmt.Sprintf("Test failed, preserving resources for debugging: %v", namespace))
+				return
+			}
+
 			err := hub.KubeClient.CoreV1().Namespaces().Delete(context.TODO(), namespace, metav1.DeleteOptions{})
 			if errors.IsNotFound(err) {
 				return
